+++ COMPARED_WITH_PRY.md
@@ -9,7 +9,7 @@ Feel free to chip in and update this table - we appreciate your help!
 | Supported Rubies          | `>= 2.0`                                                                                                             | `>= 2.7`                                                                       |                                                                                                              |
 | Source code browsing      | `show-source`                                                                                                        | `show_source`                                                                  | IRB's `show_source` can't display C source. See [#664](https://github.com/ruby/irb/issues/664)               |
 | Document browsing         | `ri`                                                                                                                 | `show_doc`                                                                     |                                                                                                              |
-| Live help system          | `help` or `command_name --help`                                                                                      | `show_cmds`                                                                    | IRB doesn't support detailed descriptions for individual commands yet                                        |
+| Live help system          | `help` or `command_name --help`                                                                                      | `help`                                                                         | IRB doesn't support detailed descriptions for individual commands yet                                        |
 | Open methods in editors   | `edit`                                                                                                               | `edit`                                                                         |                                                                                                              |
 | Syntax highlighting       | Yes                                                                                                                  | Yes                                                                            |                                                                                                              |
 | Command shell integration | Yes                                                                                                                  | No                                                                             | Currently, there's no plan to support such features in IRB                                                   |
